+++ packages/plugins/eslint-plugin-react-debug/src/rules/function-component.ts
@@ -37,7 +37,7 @@ export function create(context: RuleContext<MessageID, []>): RuleListener {
     {
       collectDisplayName: true,
       collectHookCalls: true,
-      hint: ER.DEFAULT_COMPONENT_HINT,
+      hint: ER.DEFAULT_COMPONENT_DETECTION_HINT,
     },
   );
   return {